@@ -17,10 +17,10 @@ all registered pluggable functions.
 from __future__ import annotations
 
 import logging
+from collections.abc import Callable
 from functools import wraps
 from pathlib import Path
 from types import MethodType
-from typing import Callable
 
 log = logging.getLogger(__name__)
 
